requirements.txt
streamlit>=1.36.0
google-genai>=1.0.0
pytest>=8.0.0
